+++ src/modules/_Crypto/invert.c
@@ -93,8 +93,8 @@ static void f_set_key(INT32 args)
 static void f_crypt_block(INT32 args)
 {
   char *buffer;
-  int i;
-  int len;
+  ptrdiff_t i;
+  ptrdiff_t len;
 
   if (args != 1) {
     error("Wrong number of arguments to invert->crypt_block()\n");
